refactor each component into its own file
