+++ README.md
@@ -7,10 +7,11 @@ desktop client for vex
 ## dev setup
 
 Make sure you have the dependencies:
-- g++
-- gcc
-- make
-- python (python2)
+
+-   g++
+-   gcc
+-   make
+-   python (python2)
 
 temporary icon by FreePik at FlatIcon
 window icons by Alfredo Hernandez at FlatIcon
