+++ app/components/MessagingPane.tsx
@@ -13,7 +13,6 @@ import { IconUsername } from "../components/IconUsername";
 import { selectInputStates, addInputState } from "../reducers/inputs";
 import { failMessage, ISerializedMessage } from "../reducers/messages";
 import { selectMessages } from "../reducers/messages";
-import { format } from "date-fns";
 import { markSession, selectSessions } from "../reducers/sessions";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import * as uuid from "uuid";
@@ -34,8 +33,8 @@ import { Link } from "react-router-dom";
 import { selectUser } from "../reducers/user";
 import crypto from "crypto";
 import { Highlighter } from "./Highlighter";
-import { strToIcon } from "../utils/strToIcon";
-import { allowedHighlighterTypes } from "../constants/allowedHighlighterTypes";
+import { chunkMessages } from "../utils/chunkMessages";
+import { MessageBox } from "./MessageBox";
 
 // A custom hook that builds on useLocation to parse
 // the query string for you.
@@ -588,216 +587,4 @@ export default function MessagingPane(): JSX.Element {
             </Switch>
         </div>
     );
-}
-
-export function MessageBox(props: {
-    messages: ISerializedMessage[];
-}): JSX.Element {
-    const user = useSelector(selectUser);
-    const familiars = useSelector(selectFamiliars);
-    const history = useHistory();
-
-    // don't match no characters of any length
-    const regex = /(```[^]+```)/;
-    const sender = familiars[props.messages[0]?.sender] || null;
-
-    const sessions = useSelector(selectSessions);
-
-    const sessionIDs = Object.keys(sessions[sender?.userID] || {});
-    let hasUnverifiedSession = false;
-    for (const sessionID of sessionIDs) {
-        if (!sessions[sender.userID][sessionID].verified) {
-            hasUnverifiedSession = true;
-        }
-    }
-
-    if (!sender) {
-        return <span key={uuid.v4()} />;
-    }
-
-    if (props.messages.length == 0) {
-        return <span key={crypto.randomBytes(16).toString("hex")} />;
-    }
-
-    return (
-        <article className="chat-message media" key={props.messages[0].nonce}>
-            <figure className="media-left">
-                <p className="image is-48x48">
-                    <img
-                        className="is-rounded"
-                        src={strToIcon(sender.username)}
-                    />
-                </p>
-            </figure>
-            <div className="media-content">
-                <div className="content message-wrapper">
-                    <strong>{sender.username || "Unknown User"}</strong>
-                    &nbsp;&nbsp;
-                    <small className="has-text-dark">
-                        {format(new Date(props.messages[0].timestamp), "kk:mm")}
-                    </small>
-                    &nbsp;&nbsp;
-                    {hasUnverifiedSession && sender.userID !== user.userID && (
-                        <span
-                            className="icon is-small pointer"
-                            onClick={() => {
-                                history.push(
-                                    routes.MESSAGING +
-                                        "/" +
-                                        sender.userID +
-                                        "/verify?forward=" +
-                                        history.location.pathname
-                                );
-                            }}
-                        >
-                            <FontAwesomeIcon
-                                icon={faExclamationTriangle}
-                                className={"has-text-danger"}
-                            />
-                        </span>
-                    )}
-                    <br />
-                    {props.messages.map((message) => {
-                        const isCode = regex.test(message.message);
-
-                        if (isCode) {
-                            // removing ```
-                            const languageInput = message.message
-                                .replace(/```/g, "")
-                                .split("\n")[0]
-                                .trim();
-
-                            if (
-                                allowedHighlighterTypes.includes(languageInput)
-                            ) {
-                                return (
-                                    <div
-                                        className="message-code"
-                                        key={message.nonce}
-                                    >
-                                        {Highlighter(
-                                            message.message
-                                                .replace(/```/g, "")
-                                                .replace(languageInput, "")
-                                                .trim(),
-                                            languageInput,
-                                            message.nonce
-                                        )}
-                                    </div>
-                                );
-                            }
-
-                            return (
-                                <div
-                                    className="message-code"
-                                    key={message.nonce}
-                                >
-                                    {Highlighter(
-                                        message.message
-                                            .replace(/```/g, "")
-                                            .trim(),
-                                        null
-                                    )}
-                                </div>
-                            );
-                        }
-
-                        return (
-                            <Fragment key={message.nonce}>
-                                <p className="message-text">
-                                    {!message.decrypted && (
-                                        <code>Decryption Failed</code>
-                                    )}
-                                    {message.failed ? (
-                                        <span className="has-text-danger">
-                                            {message.message}
-                                        </span>
-                                    ) : (
-                                        <span
-                                            className={`${
-                                                message.message.charAt(0) ===
-                                                ">"
-                                                    ? "has-text-success has-text-weight-bold"
-                                                    : ""
-                                            }`}
-                                        >
-                                            {message.message}
-                                        </span>
-                                    )}
-                                    &nbsp;&nbsp;
-                                    {message.failed && (
-                                        <span className="help has-text-danger">
-                                            <FontAwesomeIcon
-                                                icon={faExclamationTriangle}
-                                            />{" "}
-                                            Failed: {message.failMessage}{" "}
-                                        </span>
-                                    )}
-                                </p>
-                            </Fragment>
-                        );
-                    })}
-                </div>
-            </div>
-            <div className="media-right" />
-        </article>
-    );
-}
-
-/**
- * Chunks the messages into chunks of the same sender consecutively, so they can be displayed in
- * a smaller format.
- *
- *
- *
- * @param threadMessages The thread message object.
- */
-export function chunkMessages(
-    threadMessages: Record<string, ISerializedMessage>
-): ISerializedMessage[][] {
-    const messageIDs = Object.keys(threadMessages);
-
-    const chunkedMessages: ISerializedMessage[][] = [[]];
-    for (let i = 0; i < messageIDs.length; i++) {
-        if (chunkedMessages[0] === undefined) {
-            chunkedMessages.push([]);
-        }
-
-        const currentMessage = threadMessages[messageIDs[i]];
-
-        if (chunkedMessages[chunkedMessages.length - 1].length === 0) {
-            chunkedMessages[chunkedMessages.length - 1].push(currentMessage);
-        } else {
-            if (
-                chunkedMessages[chunkedMessages.length - 1][0].sender ===
-                currentMessage.sender
-            ) {
-                const firstMessageTime = new Date(
-                    chunkedMessages[chunkedMessages.length - 1][0].timestamp
-                );
-                const thisMessageTime = new Date(currentMessage.timestamp);
-                const chunkTimeLength =
-                    thisMessageTime.getTime() - firstMessageTime.getTime();
-
-                // three minutes
-                if (chunkTimeLength < 1000 * 60 * 3) {
-                    chunkedMessages[chunkedMessages.length - 1].push(
-                        currentMessage
-                    );
-                    // start a new chunk if the chunk is too old
-                } else {
-                    chunkedMessages.push([]);
-                    chunkedMessages[chunkedMessages.length - 1].push(
-                        currentMessage
-                    );
-                }
-            } else {
-                chunkedMessages.push([]);
-                chunkedMessages[chunkedMessages.length - 1].push(
-                    currentMessage
-                );
-            }
-        }
-    }
-    return chunkedMessages;
 }
