+++ app/views/Messaging.tsx
@@ -1,10 +1,11 @@
 import React, { Fragment } from "react";
-import MessagingBar, { UserMenu } from "../components/MessagingBar";
+import MessagingBar from "../components/MessagingBar";
 import MessagingPane from "../components/MessagingPane";
 import { useSelector } from "react-redux";
 import { selectApp } from "../reducers/app";
 import Loading from "../components/Loading";
 import { ServerBar } from "../components/ServerBar";
+import { UserMenu } from "../components/UserMenu";
 
 export default function Messaging(): JSX.Element {
     const app = useSelector(selectApp);
