+++ app/views/Register.tsx
@@ -4,33 +4,19 @@ import { Client } from "@vex-chat/vex";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { addInputState, selectInputStates } from "../reducers/inputs";
-import { errorFX, gaurdian, switchFX, VerticalAligner } from "../views/Base";
+import { errorFX, gaurdian, switchFX } from "../views/Base";
 import { useHistory } from "react-router";
 import { dbFolder, keyFolder } from "../components/ClientLauncher";
 import { routes } from "../constants/routes";
 import { loadKeyFile, saveKeyFile } from "../utils/KeyGaurdian";
+import { BackButton } from "../components/BackButton";
+import { VerticalAligner } from "../components/VerticalAligner";
 
 const USERNAME_INPUT_NAME = "register_username";
 const PASSWORD_INPUT_NAME = "register_password";
 const CONFIRM_INPUT_NAME = "register_confirmpass";
 
-export const backButton = (route?: string): JSX.Element => {
-    const history = useHistory();
-    return (
-        <a
-            className="delete settings-delete is-medium"
-            onClick={() => {
-                if (route) {
-                    history.push(route);
-                } else {
-                    history.goBack();
-                }
-            }}
-        />
-    );
-};
-
-export default function IRegister(): JSX.Element {
+export default function Register(): JSX.Element {
     const dispatch = useDispatch();
     const history = useHistory();
 
@@ -111,7 +97,7 @@ export default function IRegister(): JSX.Element {
     };
 
     return (
-        <VerticalAligner top={backButton(routes.HOME)}>
+        <VerticalAligner top={<BackButton route={routes.HOME} />}>
             <div className="box has-background-white register-form">
                 <div className="field">
                     <label className="label is-small">
