+++ app/views/Settings.tsx
@@ -1,6 +1,8 @@
 import React, { useState } from "react";
-import { dataStore, VerticalAligner } from "./Base";
-import { backButton } from "./Register";
+import { BackButton } from "../components/BackButton";
+import { VerticalAligner } from "../components/VerticalAligner";
+import { routes } from "../constants/routes";
+import { dataStore } from "./Base";
 
 export default function Settings(): JSX.Element {
     const [notification, setNotifications] = useState(
@@ -8,7 +10,7 @@ export default function Settings(): JSX.Element {
     );
 
     return (
-        <VerticalAligner top={backButton()}>
+        <VerticalAligner top={<BackButton route={routes.HOME} />}>
             <div className="panel is-light">
                 <p className="panel-heading">Settings</p>
                 <div className="panel-block">
